move files under `internal` directory
