+++ internal/cmd/event.go
