+++ internal/cmd/eventDelete.go
