+++ internal/cmd/eventRemoveTag.go
@@ -1,16 +1,17 @@
 package cmd
 
 import (
-	"kakeibodb/mysql_client"
-	"kakeibodb/usecase"
 	"log"
+
+	"kakeibodb/internal/mysql_client"
+	"kakeibodb/internal/usecase"
 
 	"github.com/spf13/cobra"
 )
 
-// eventAddTagCmd represents the addTag command
-var eventAddTagCmd = &cobra.Command{
-	Use:   "addTag",
+// eventRemoveTagCmd represents the removeTag command
+var eventRemoveTagCmd = &cobra.Command{
+	Use:   "removeTag",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -23,32 +24,32 @@ to quickly create a Cobra application.`,
 		if err != nil {
 			log.Fatal(err)
 		}
-		tagNames, err := cmd.Flags().GetStringSlice("tagNames")
+		tagName, err := cmd.Flags().GetString("tagName")
 		if err != nil {
 			log.Fatal(err)
 		}
-		if eventID == 0 && len(tagNames) == 0 {
-			log.Fatal("both eventID and tagNames must be specified.")
+		if eventID == 0 && tagName == "" {
+			log.Fatal("both eventID and tagName must be specified.")
 		}
 
 		eh := usecase.NewEventHandler(mysql_client.NewMySQLClient(dbName, user))
 		defer eh.Close()
-		eh.AddTag(eventID, tagNames)
+		eh.RemoveTag(eventID, tagName)
 	},
 }
 
 func init() {
-	eventCmd.AddCommand(eventAddTagCmd)
+	eventCmd.AddCommand(eventRemoveTagCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// addTagCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// removeTagCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// addTagCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-	eventAddTagCmd.Flags().Int("eventID", 0, "credit == false: Event ID, credit == true: Credit card event ID")
-	eventAddTagCmd.Flags().StringSlice("tagNames", nil, "Tag Names")
+	// removeTagCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	eventRemoveTagCmd.Flags().Int("eventID", 0, "Event ID")
+	eventRemoveTagCmd.Flags().StringP("tagName", "t", "", "Tag Name")
 }
