+++ internal/cmd/pattern.go
