+++ internal/cmd/patternList.go
@@ -1,16 +1,15 @@
 package cmd
 
 import (
-	"kakeibodb/mysql_client"
-	"kakeibodb/usecase"
-	"log"
+	"kakeibodb/internal/mysql_client"
+	"kakeibodb/internal/usecase"
 
 	"github.com/spf13/cobra"
 )
 
-// tagDeleteCmd represents the tagDelete command
-var tagDeleteCmd = &cobra.Command{
-	Use:   "delete",
+// patternListCmd represents the list command
+var patternListCmd = &cobra.Command{
+	Use:   "list",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
 and usage of using your command. For example:
@@ -19,31 +18,22 @@ Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		tagID, err := cmd.Flags().GetInt("tagID")
-		if err != nil {
-			log.Fatal(err)
-		}
-		if tagID == 0 {
-			log.Fatal("tagID must be specified.")
-		}
-
-		th := usecase.NewTagHandler(mysql_client.NewMySQLClient(dbName, user))
-		defer th.Close()
-		th.DeleteTag(tagID)
+		lh := usecase.NewListHandler(mysql_client.NewMySQLClient(dbName, user))
+		defer lh.Close()
+		lh.ListAllPattern()
 	},
 }
 
 func init() {
-	tagCmd.AddCommand(tagDeleteCmd)
+	patternCmd.AddCommand(patternListCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// tagDeleteCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// listCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// tagDeleteCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-	tagDeleteCmd.Flags().Int("tagID", 0, "Tag ID")
+	// listCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
