+++ internal/cmd/tagDelete.go
@@ -1,15 +1,16 @@
 package cmd
 
 import (
-	"kakeibodb/mysql_client"
-	"kakeibodb/usecase"
 	"log"
+
+	"kakeibodb/internal/mysql_client"
+	"kakeibodb/internal/usecase"
 
 	"github.com/spf13/cobra"
 )
 
-// patternDeleteCmd represents the delete command
-var patternDeleteCmd = &cobra.Command{
+// tagDeleteCmd represents the tagDelete command
+var tagDeleteCmd = &cobra.Command{
 	Use:   "delete",
 	Short: "A brief description of your command",
 	Long: `A longer description that spans multiple lines and likely contains examples
@@ -19,33 +20,31 @@ Cobra is a CLI library for Go that empowers applications.
 This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		patternID, err := cmd.Flags().GetInt("patternID")
+		tagID, err := cmd.Flags().GetInt("tagID")
 		if err != nil {
 			log.Fatal(err)
 		}
-		if patternID == 0 {
-			log.Fatal("patternID must be specified.")
+		if tagID == 0 {
+			log.Fatal("tagID must be specified.")
 		}
 
-		ph := usecase.NewPatternHandler(mysql_client.NewMySQLClient(dbName, user))
-		defer ph.Close()
-		ph.DeletePattern(patternID)
+		th := usecase.NewTagHandler(mysql_client.NewMySQLClient(dbName, user))
+		defer th.Close()
+		th.DeleteTag(tagID)
 	},
 }
 
 func init() {
-	patternCmd.AddCommand(patternDeleteCmd)
+	tagCmd.AddCommand(tagDeleteCmd)
 
 	// Here you will define your flags and configuration settings.
 
 	// Cobra supports Persistent Flags which will work for this command
 	// and all subcommands, e.g.:
-	// deleteCmd.PersistentFlags().String("foo", "", "A help for foo")
+	// tagDeleteCmd.PersistentFlags().String("foo", "", "A help for foo")
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// deleteCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
-	patternDeleteCmd.Flags().Int("patternID", 0, "Pattern ID")
-
-	patternDeleteCmd.MarkFlagRequired("patternID")
+	// tagDeleteCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	tagDeleteCmd.Flags().Int("tagID", 0, "Tag ID")
 }
