+++ internal/usecase/loadEvent.go
@@ -2,12 +2,13 @@ package usecase
 
 import (
 	"fmt"
-	"kakeibodb/db_client"
-	"kakeibodb/event"
 	"log"
 	"os"
 	"path/filepath"
 	"strconv"
+
+	"kakeibodb/internal/db_client"
+	"kakeibodb/internal/event"
 )
 
 type LoadEventHandler struct {
